% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_oxy_lps_pellet}
\alias{data_oxy_lps_pellet}
\title{RAW 264.7 Macrophage Pellet Oxylipin Data (LPS Stimulated)}
\format{
A data frame with oxylipin species as rows and 8 samples as columns.
Abundances are normalized to total protein content (fmol/mg protein).
}
\source{
Internal experimental dataset following the standardized
oxylipidomics protocol.
}
\usage{
data_oxy_lps_pellet
}
\description{
Intracellular oxylipin abundance profiles from RAW 264.7 murine
macrophages. This dataset compares basal inflammatory states (Control)
against activation with Lipopolysaccharide (LPS, 100 ng/mL).
}
\keyword{datasets}
