% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_oxy_lps_supernatant}
\alias{data_oxy_lps_supernatant}
\title{RAW 264.7 Macrophage Supernatant Oxylipin Data (LPS Stimulated)}
\format{
A data frame with oxylipin species as rows and 7 samples as columns.
Abundances are normalized by sample volume (fmol/mL).
}
\source{
Internal experimental dataset following the standardized
oxylipidomics protocol.
}
\usage{
data_oxy_lps_supernatant
}
\description{
Secreted oxylipin signatures (exometabolome) from RAW 264.7 murine
macrophages. It captures the flux of lipid mediators released into the
extracellular environment following TLR4 activation.
}
\keyword{datasets}
