% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_oxy_pellet}
\alias{data_oxy_pellet}
\title{Oxylipidomics Abundance Data (Cellular Pellets)}
\format{
A data frame with lipids as rows and samples as columns.
}
\source{
Internal experimental data from RAW 264.7 cell line studies.
}
\usage{
data_oxy_pellet
}
\description{
A dataset containing oxylipidomics abundance levels from RAW 264.7 murine
macrophage pellets. This dataset compares two harvesting methods:
mechanical scraping (Scraper, n = 6) and enzymatic treatment (Trypsin, n = 6).
}
\keyword{datasets}
