% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_oxy_supernatant}
\alias{data_oxy_supernatant}
\title{Oxylipidomics Abundance Data (Cell Culture Supernatant)}
\format{
A data frame with lipids as rows and samples as columns.
}
\source{
Secreted inflammatory mediator profiles in macrophage supernatants.
}
\usage{
data_oxy_supernatant
}
\description{
A dataset of secreted oxylipins in RAW 264.7 culture media. It evaluates
the impact of harvesting procedures across three conditions: Scraper (n = 6),
Trypsin (n = 6), and Control (no manipulation, n = 6).
}
\keyword{datasets}
