% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analyzed_oxy.R
\name{get_analyzed_oxy}
\alias{get_analyzed_oxy}
\title{Retrieve Names of Analyzed Oxylipins}
\usage{
get_analyzed_oxy(obj)
}
\arguments{
\item{obj}{A \code{staRoxy} object.}
}
\value{
A character vector containing the names of the lipids (row names
of the data matrix).
}
\description{
A simple helper function to extract the names of all oxylipin features
currently retained in the dataset. This reflects the state of the data
after any filtering or subsetting steps.
}
\examples{
\dontshow{
staRoxy_object <- read_oxy(data_oxy_lps_pellet, metadata_oxy_lps_pellet)
staRoxy_object <- filter_oxy(staRoxy_object)
staRoxy_object <- transform_oxy(staRoxy_object)
}

# Retrieve the names of the oxylipins currently in the object
get_analyzed_oxy(staRoxy_object)

}
