% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oxy_universe.R
\name{get_oxy_universe}
\alias{get_oxy_universe}
\title{Get Default Oxylipin Reference Universe}
\usage{
get_oxy_universe()
}
\value{
A data frame with two columns: \code{Oxylipin} and \code{Precursor}.
}
\description{
Retrieves the default \code{staRoxy} reference universe, mapping oxylipins
to their fatty acid precursors. This object is used natively by the
\code{oxy_ora} function.
}
\examples{
get_oxy_universe()

}
