% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_contrasts.R
\name{list_contrasts}
\alias{list_contrasts}
\title{List Experimental Contrasts}
\usage{
list_contrasts(stats_obj)
}
\arguments{
\item{stats_obj}{The list object returned by \code{limma_oxy}.}
}
\value{
Invisibly returns the contrast names. Prints to console.
}
\description{
Helper function to display all pairwise comparisons available in a
\code{limma_oxy} results object.
}
\examples{
\dontshow{
staRoxy_object <- read_oxy(data_oxy_lps_pellet, metadata_oxy_lps_pellet)
staRoxy_object <- filter_oxy(staRoxy_object)
staRoxy_object <- transform_oxy(staRoxy_object)
}

# List all pairwise comparisons generated by limma_oxy
stats <- limma_oxy(staRoxy_object, save_results = FALSE)

list_contrasts(stats)

}
