% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_oxy_lps_supernatant}
\alias{metadata_oxy_lps_supernatant}
\title{Metadata for RAW 264.7 Supernatant Samples}
\format{
A data frame with 7 rows and 2 columns:
\describe{
\item{sample}{Unique sample identifier matching data headers}
\item{group}{Experimental group (Control or LPS)}
}
}
\usage{
metadata_oxy_lps_supernatant
}
\description{
Experimental metadata mapping the 8 culture supernatant samples to their
respective groups: "Control" or "LPS".
}
\keyword{datasets}
