% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_oxy_pellet}
\alias{metadata_oxy_pellet}
\title{Metadata for Cellular Pellet Samples}
\format{
A data frame with 12 rows and 2 columns:
\describe{
\item{sample}{Unique sample identifier}
\item{group}{Harvesting method used (Scraper or Trypsin)}
}
}
\usage{
metadata_oxy_pellet
}
\description{
Mapping of sample identifiers to experimental groups (Scraper vs. Trypsin)
for the cellular pellet dataset.
}
\keyword{datasets}
