% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_oxy_plasma}
\alias{metadata_oxy_plasma}
\title{Metadata for Blood Plasma Samples}
\format{
A data frame with 24 rows and 2 columns:
\describe{
\item{sample}{Unique sample identifier matching data headers}
\item{group}{Experimental group: ALS_70d, WT_70d, ALS_120d, or WT_120d}
}
}
\source{
Chaves-Filho, A. B., et al. (2023). \doi{10.1016/j.freeradbiomed.2023.08.019}
}
\usage{
metadata_oxy_plasma
}
\description{
Mapping of 24 sample identifiers to their respective experimental groups
within the ALS mouse model study.
}
\keyword{datasets}
