% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_na_diagnostic.R
\name{plot_na_diagnostic}
\alias{plot_na_diagnostic}
\title{Missing Data Diagnostic for Oxylipins}
\usage{
plot_na_diagnostic(
  obj,
  mode = "cor",
  point_color = "#2C56A0",
  true_color = "#C9273E",
  false_color = "#57B894",
  line_color = "black",
  title_size = 14,
  x_axis_size = 12,
  y_axis_size = 12,
  legend_title_size = 12,
  legend_size = 12
)
}
\arguments{
\item{obj}{A \code{staRoxy} object.}

\item{mode}{Character. Either \code{"cor"} (default) or \code{"dist"}.}

\item{point_color, true_color, false_color, line_color}{Character. Hex codes or names for plot colors.}

\item{title_size, x_axis_size, y_axis_size, legend_title_size, legend_size}{Numeric. Font sizes for plot elements.}
}
\value{
A \code{ggplot2} object (mode \code{"cor"}) or a \code{cowplot} grid object (mode \code{"dist"}).
}
\description{
Visualizes missing value patterns to distinguish between Missing Not At Random
(MNAR), typically associated with the Limit of Detection (LOD), and
Missing At Random (MAR), arising from experimental or technical variability.
}
\details{
The function provides two diagnostic modes:
\itemize{
\item \strong{Correlation mode ("cor"):} Generates a scatter plot of Mean
Log2 Abundance vs. Proportion of Missing Values. A negative correlation
suggests an MNAR pattern, where lipids are more likely to be missing
at lower abundances.
\item \strong{Distribution mode ("dist"):} Compares the abundance
distributions of features with and without missing values using Density
and ECDF plots. A leftward shift for features with NAs (\code{TRUE})
is a strong indicator of MNAR.
}

\strong{Implication for Imputation:}
\itemize{
\item If MNAR is detected, use \code{na_method = "minprob"} in downstream functions.
\item If MAR is detected, use \code{na_method = "rf"} (Random Forest).
}
}
\examples{
\dontshow{
staRoxy_object <- read_oxy(data_oxy_lps_pellet, metadata_oxy_lps_pellet)
staRoxy_object <- filter_oxy(staRoxy_object)
staRoxy_object <- transform_oxy(staRoxy_object)
}

\dontrun{
# Diagnostic using correlation mode (default)
# Ideal to see if lower abundance correlates with more missing values
plot_na_diagnostic(staRoxy_object, mode = "cor")

# Diagnostic using distribution mode
# Ideal to compare density and cumulative distribution shifts
plot_na_diagnostic(staRoxy_object, mode = "dist")
}

}
