% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rank_oxy.R
\name{plot_rank_oxy}
\alias{plot_rank_oxy}
\title{Plot Ranked Differential Abundance}
\usage{
plot_rank_oxy(
  stats,
  contrast = 1,
  top_n = 10,
  rank_by = "logFC",
  palette = "magma",
  direction = -1,
  title_size = 14,
  x_axis_size = 12,
  y_axis_size = 12,
  legend_title_size = 12,
  legend_size = 12,
  sig_size = 6
)
}
\arguments{
\item{stats}{The list object returned by \code{limma_oxy}.}

\item{contrast}{Numeric (index) or Character (name) of the contrast to plot.}

\item{top_n}{Integer. Number of oxylipins to display.}

\item{rank_by}{Character. Rank by \code{"logFC"} (absolute) or \code{"p"} (significance).}

\item{palette}{Character. Viridis palette option (e.g., "magma", "viridis").}

\item{direction}{Integer. Direction of the color scale (1 or -1).}

\item{title_size, x_axis_size, y_axis_size, legend_title_size, legend_size, sig_size}{Numeric. Font sizes for plot elements and significance asterisks.}
}
\value{
A ggplot2 object.
}
\description{
Visualizes the top differentially abundant oxylipins using a horizontal
bar plot. Features are ranked by either Fold-Change or Significance.
}
\details{
Standard Error of the Mean (SEM) is estimated as \code{abs(logFC / t)}.
Significance levels are indicated by asterisks based on adjusted P-values.
}
\examples{
\dontshow{
staRoxy_object <- read_oxy(data_oxy_lps_pellet, metadata_oxy_lps_pellet)
staRoxy_object <- filter_oxy(staRoxy_object)
staRoxy_object <- transform_oxy(staRoxy_object)
}

\dontrun{
# Plot the top 10 features for the first contrast, ranked by absolute LogFC
stats <- limma_oxy(staRoxy_object, save_results = FALSE)

plot_rank_oxy(stats, contrast = 1, top_n = 10, rank_by = "logFC")
}

}
