% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_oxy_plasma}
\alias{data_oxy_plasma}
\title{Oxylipidomics Abundance Data (Blood Plasma - ALS Mouse Model)}
\format{
A data frame with lipids as rows and 24 samples as columns.
}
\source{
Chaves-Filho, A. B., et al. (2023). Free Radical Biology and Medicine.
\doi{10.1016/j.freeradbiomed.2023.08.019}
}
\usage{
data_oxy_plasma
}
\description{
A comprehensive oxylipidomics dataset from blood plasma of SOD1-G93A mice
(a model for Amyotrophic Lateral Sclerosis) and age-matched wild-type (WT)
controls. Samples were collected at two biological stages: asymptomatic
(70 days old) and symptomatic (120 days old).
}
\keyword{datasets}
