% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analyzed_oxy.R
\name{get_analyzed_oxy}
\alias{get_analyzed_oxy}
\title{Retrieve Names of Analyzed Oxylipins}
\usage{
get_analyzed_oxy(obj)
}
\arguments{
\item{obj}{A \code{staRoxy} object.}
}
\value{
A character vector containing the names of the lipids (row names
of the data matrix).
}
\description{
A simple helper function to extract the names of all oxylipin features
currently retained in the dataset. This reflects the state of the data
after any filtering or subsetting steps.
}
\examples{
# List all lipids currently in the pellet dataset
get_analyzed_oxy(data_oxy_pellet)

}
