% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_contrasts.R
\name{list_contrasts}
\alias{list_contrasts}
\title{List Experimental Contrasts}
\usage{
list_contrasts(stats_obj)
}
\arguments{
\item{stats_obj}{The list object returned by \code{limma_oxy}.}
}
\value{
Invisibly returns the contrast names. Prints to console.
}
\description{
Helper function to display all pairwise comparisons available in a
\code{limma_oxy} results object.
}
